OpenTracker.UnitTests/Models/UndoRedo/Mode/ChangeCompassShuffleTests.cs
using Autofac;
using NSubstitute;
using OpenTracker.Models.Modes;
using OpenTracker.Models.UndoRedo.Mode;
using Xunit;

namespace OpenTracker.UnitTests.Models.UndoRedo.Mode;

public class ChangeCompassShuffleTests
{
    private readonly IMode _mode = Substitute.For<IMode>();

    [Fact]
    public void CanExecute_ShouldReturnTrueAlways()
    {
        var sut = new ChangeCompassShuffle(_mode, false);
            
        Assert.True(sut.CanExecute());
    }

    [Theory]
    [InlineData(false, false)]
    [InlineData(true, true)]
    public void ExecuteDo_ShouldSetCompassShuffleToNewValue(bool expected, bool newValue)
    {
        var sut = new ChangeCompassShuffle(_mode, newValue);
        sut.ExecuteDo();
            
        Assert.Equal(expected, _mode.CompassShuffle);
    }

    [Theory]
    [InlineData(false, false)]
    [InlineData(true, true)]
    public void ExecuteUndo_ShouldSetCompassShuffleToPreviousValue(bool expected, bool previousValue)
    {
        _mode.CompassShuffle.Returns(previousValue);
        var sut = new ChangeCompassShuffle(_mode, !previousValue);
        sut.ExecuteDo();
        sut.ExecuteUndo();
            
        Assert.Equal(expected, _mode.CompassShuffle);
    }

    [Fact]
    public void AutofacTest()
    {
        using var scope = ContainerConfig.Configure().BeginLifetimeScope();
        var factory = scope.Resolve<IChangeCompassShuffle.Factory>();
        var sut = factory(false);
            
        Assert.NotNull(sut as ChangeCompassShuffle);
    }
}